Move dataStrategys test to their own file
